Extract installShadowStyles into #core
