+++ src/core/dom/web-components.js
@@ -1,3 +1,5 @@
+import {toWin} from '#core/window';
+
 /**
  * Possible versions of Shadow DOM spec
  * @enum {string}
@@ -96,4 +98,33 @@ function getShadowDomVersion(element) {
     return ShadowDomVersion_Enum.V0;
   }
   return ShadowDomVersion_Enum.NONE;
+}
+
+/**
+ * @param {ShadowRoot} shadowRoot
+ * @param {string} name
+ * @param {string} cssText
+ */
+export function installShadowStyle(shadowRoot, name, cssText) {
+  const doc = shadowRoot.ownerDocument;
+  const win = toWin(doc.defaultView);
+  if (
+    shadowRoot.adoptedStyleSheets !== undefined &&
+    win.CSSStyleSheet.prototype.replaceSync !== undefined
+  ) {
+    const cache = win[SHADOW_CSS_CACHE] || (win[SHADOW_CSS_CACHE] = {});
+    let styleSheet = cache[name];
+    if (!styleSheet) {
+      styleSheet = new win.CSSStyleSheet();
+      styleSheet.replaceSync(cssText);
+      cache[name] = styleSheet;
+    }
+    shadowRoot.adoptedStyleSheets =
+      shadowRoot.adoptedStyleSheets.concat(styleSheet);
+  } else {
+    const styleEl = doc.createElement('style');
+    styleEl.setAttribute('data-name', name);
+    styleEl.textContent = cssText;
+    shadowRoot.appendChild(styleEl);
+  }
 }
