+++ src/preact/base-element.js
@@ -20,6 +20,7 @@ import {
 import {MediaQueryProps} from '#core/dom/media-query-props';
 import {childElementByAttr, childElementByTag} from '#core/dom/query';
 import {PauseHelper} from '#core/dom/video/pause-helper';
+import {installShadowStyle} from '#core/dom/web-components';
 import * as mode from '#core/mode';
 import {isElement} from '#core/types';
 import {dict, hasOwn, map} from '#core/types/object';
@@ -36,8 +37,6 @@ import {
   checkPropsFor,
   collectProps,
 } from './parse-props';
-
-import {installShadowStyle} from '../shadow-embed';
 
 /** @const {MutationObserverInit} */
 const CHILDREN_MUTATION_INIT = {
